refactor(cli): Introduce 'platform' command group for status
Move `prisma status` command under a new `prisma platform` command group. This refactoring aims to organize CLI commands related to the Prisma Data Platform.

The changes include:
- Extracting status fetching and formatting logic into a dedicated `status-page.ts` module.
- Introducing new modules (`platform/$`, `platform/_lib/cli/dispatchToSubCommand`, `platform/_lib/help`) for handling the `platform` command group and its subcommands.
- Updating the main CLI entry point (`bin.ts`) to register the new `platform` command.
- Refactoring the `Status` command to delegate to the new `fetchStatus` function and adjusting tests accordingly.
- Updating help texts to reflect the new command structure.
