+++ packages/cli/src/CLI.ts
@@ -139,7 +139,7 @@ Learn more at ${link('https://pris.ly/cli/pdp')}`
               format   Format your Prisma schema
              version   Displays Prisma version info
                debug   Displays Prisma debug info
-              status   Show Prisma Data Platform service status
+            platform   Prisma Data Platform commands
                  mcp   Starts an MCP server to use with AI development tools
 
     ${bold('Flags')}
