+++ packages/cli/src/platform/_.ts
@@ -1,0 +1,1 @@
+export * from './$'
