+++ packages/cli/src/platform/_Platform.ts
@@ -1,0 +1,1 @@
+export * as Platform from './_'
